Fix gcds-menu-group appearing as a dependency in all components
