+++ packages/web/src/components/gcds-date-modified/readme.md
@@ -5,20 +5,6 @@
 <!-- Auto Generated Below -->
 
 
-## Dependencies
-
-### Depends on
-
-- [gcds-menu-group](../gcds-menu-group)
-
-### Graph
-```mermaid
-graph TD;
-  gcds-date-modified --> gcds-menu-group
-  gcds-menu-group --> gcds-menu-group
-  style gcds-date-modified fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
