+++ packages/web/src/components/gcds-error-summary/readme.md
@@ -15,20 +15,6 @@
 | `subHeading` | `sub-heading` | Set error summary message                                                                | `string`           | `undefined` |
 
 
-## Dependencies
-
-### Depends on
-
-- [gcds-menu-group](../gcds-menu-group)
-
-### Graph
-```mermaid
-graph TD;
-  gcds-error-summary --> gcds-menu-group
-  gcds-menu-group --> gcds-menu-group
-  style gcds-error-summary fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
