+++ packages/web/src/components/gcds-stepper/readme.md
@@ -13,20 +13,6 @@
 | `totalSteps` _(required)_  | `total-steps`  | Defines the total amount of steps. | `number` | `undefined` |
 
 
-## Dependencies
-
-### Depends on
-
-- [gcds-menu-group](../gcds-menu-group)
-
-### Graph
-```mermaid
-graph TD;
-  gcds-stepper --> gcds-menu-group
-  gcds-menu-group --> gcds-menu-group
-  style gcds-stepper fill:#f9f,stroke:#333,stroke-width:4px
-```
-
 ----------------------------------------------
 
 *Built with [StencilJS](https://stenciljs.com/)*
